Restructure Architecture
