+++ src/resources/schemas/presentation.json
@@ -1,0 +1,35 @@
+{
+    "type": "object",
+    "properties": {
+        "total": {
+            "type": "integer"
+        },
+        "list": {
+            "type": "array",
+            "items": {
+                "type": "object",
+                "properties": {
+                    "id": {"type": "string"},
+                    "name": {"type": "string"},
+                    "status": {"type": "string"},
+                    "dateStart": {"type": "string"},
+                    "dateEnd": {"type": "string"},
+                    "dateStartDate": {"type": ["null", "string"]},
+                    "dateEndDate": {"type": ["null", "string"]},
+                    "parentId": {"type": ["string", "null"]},
+                    "parentType": {"type": ["string", "null"]},
+                    "parentName": {"type": ["string", "null"]},
+                    "createdById": {"type": "string"},
+                    "assignedUserId": {"type": "string"},
+                    "assignedUserName": {"type": "string"}
+                },
+                "required": [
+                    "id", "name", "status", "dateStart", "dateEnd",
+                    "dateStartDate", "dateEndDate", "parentId", "parentType",
+                    "parentName", "createdById", "assignedUserId", "assignedUserName"
+                ]
+            }
+        }
+    },
+    "required": ["total", "list"]
+}
